requirements.txt
jsonpickle
botbuilder-integration-aiohttp>=4.14.0
redis
openai
requests
azure-cognitiveservices-search-websearch
azure-identity
azure-search
azure-search-documents
azure-core
azure-data-tables
azure-cosmosdb-table
aiohttp-cors
adaptivecards